only superusers can access add, delete, and edit products

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -89,16 +90,20 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     # Handles adding a new product to the database manually or
     # by searching for it on untappd
     """
+    if not request.user.is_superuser:
+        # If user is not an admin, redirect to home page
+        messages.error(request, "Only store owners can access this page.")
+        return redirect(reverse('home'))
+
     if request.POST:
         if 'product_search' in request.POST:
-            """
             # If product search on untappd form has been submitted
-            """
             search_query = request.POST['product_search']
             results = UntappdHandler.search_beer(search_query)
 
@@ -115,9 +120,7 @@ def add_product(request):
                 }
 
         elif 'beer_id' in request.POST:
-            """
             # If product search result has been selected
-            """
             beer_id = request.POST['beer_id']
             beer_info = UntappdHandler.get_beer_info(beer_id)
 
@@ -170,11 +173,18 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     # Edits a product in the store
     """
+    if not request.user.is_superuser:
+        # If user is not an admin, redirect to home page
+        messages.error(request, "Only store owners can access this page.")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
         if form.is_valid():
@@ -196,10 +206,16 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     # Delete a product from the store
     """
+    if not request.user.is_superuser:
+        # If user is not an admin, redirect to home page
+        messages.error(request, "Only store owners can access this page.")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Successfully removed product from the shop.')
